fixup! JAMES-3693 Implement RedisExtension for testing
